# Optimize bias_add reluv2 in half2