<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
Performance optimization
### PR changes
OPs
### Describe
之前的bias_add_relu在half情况下只做了half2的向量化。但向量化最大支持128bit，即使是half2也可以打包成4个half2来做向量化。这个PR实现了4/2/1 三种向量化大小，并进行dispatch

取TCIR中的数据测试：

1. row = 132480, col = 360

|  Device  | Before | After |
| ---- | ---- | ---- |
|  A10  |  375us |  388us | 
|  A100  |  150us |  132.9us | 

2. row = 132480, col = 540

|  Device  | Before | After |
| ---- | ---- | ---- |
|  A10  |  us |  us | 
|  A100  |  225us |  199us | 

这是一个memory bound算子，所以优势在低带宽机器上不明显